requirements.txt
python-dotenv
discord
discord.py
groq
